Integrated Speech recognition
The basic functions zooming in/out, panning left/r/u/d will be
recognized, but for now nothing happens on the map. In the chrome
console you can see what the tool recognizes while u speak to it. Would
be great if some1 could trigger the map behavior inside the functions or
if we work on it together in a meeting.

Since the webspeech api in some cases missinterpretes e.g. a "right" as
a "white", i suggest to trigger the command for panning right also on
slightly wrong missinterpreted words such as "white" instead of "right".

tutorial link:
http://stiltsoft.com/blog/2013/05/google-chrome-how-to-use-the-web-speech-api/